fix(wash-runtime): keep a metered host serving HTTP, and stop one insecure registry from breaking every other
A host started with `--enable-meters` — which the chart's hostgroup passes by
default — served zero HTTP. Every request came back 500 with an empty body and
no guest logs, because the guest never ran.

A store built on a fuel-enabled engine starts at **zero** fuel, and
instantiation runs guest code, so the component trapped on instantiate before
`FuelConsumptionMeter::observe` ever got to set a budget. The NATS subscriber
had already learned this and primed its own stores; every other path had not.
Fixed where stores are made — `new_store_from_templates` and the component-host
plugin store — rather than at each call site, so P2 HTTP, P3 HTTP, messaging,
and the plugins are all covered by one line. Fuel here is metering, never a
limit: the epoch deadline is what ends a guest that will not stop.

Three of the reasons that took so long to find are fixed alongside it:

- The invoke error path threw away the guest's real error. `task.await` yields
  `Result<anyhow::Result<()>, JoinError>` and only the `JoinError` arm was
  matched, so every guest-side failure — instantiation, metering, `call_handle`
  — surfaced as the same "oneshot channel closed but no response was sent".
  Both levels are matched now, and `Ok(Err(e))` is propagated.
- Two unrelated workloads that both claim one HTTP hostname split each other's
  requests at random, with nothing logged. Registration now carries the
  workload's `<namespace>/<name>`, so replicas of one deployment still share a
  hostname silently while two different deployments warn.
- `--allow-insecure-registries` forced plain HTTP on *every* registry, so
  enabling it for a local kind registry broke `ghcr.io` with `Not authorized` —
  a scheme problem that reads as a credentials problem. Added
  `--insecure-registry=<host[:port]>` (repeatable,
  `WASH_INSECURE_REGISTRIES`), which names the registries that may use HTTP and
  leaves the rest on HTTPS; the blunt flag still means what it always did and
  now warns at startup.

`integration_http_fuel_metering` covers the regression: a host with metering
on, serving a P2 component, a P3 component, and a plugin-using one. Reverting
the priming line fails all three with the reported symptom — 500, empty body.
