+++ crates/wash-runtime/src/engine/linked_call.rs
@@ -353,6 +353,14 @@ pub(crate) async fn new_store_from_templates(
     }
 
     let mut store = wasmtime::Store::new(engine, shared_ctx);
+    // Fuel here is *metering*, not a limit: `FuelConsumptionMeter::observe`
+    // measures a call by the fuel it burns, and the epoch deadline below — not
+    // fuel — is what ends a guest that will not stop. But a store built on a
+    // fuel-enabled engine starts at **zero** fuel, and instantiation runs guest
+    // code, so without a budget here every component on a host started with
+    // `--enable-meters` traps before any metering wrapper is reached. Errors
+    // when the engine has fuel off, which is not a failure.
+    let _ = store.set_fuel(u64::MAX);
     // Trap for every store built here, services included: trapping a service
     // means a supervisor restart, which beats carrying a wedged call forever.
     arm_epoch_deadline(&mut store, AbandonedCallPolicy::Trap);
