--- crates/wash-runtime/tests/fixtures/nats-core-handler/wit/world.wit
@@ -1,7 +1,0 @@
-package wasmcloud:nats-core-handler-fixture;
-
-world nats-core-handler-fixture {
-    import wasmcloud:nats/types@0.1.0;
-    import wasmcloud:nats/core@0.1.0;
-    export wasmcloud:nats/core-handler@0.1.0;
-}
