--- crates/wash-runtime/tests/fixtures/nats-jetstream-handler/.wash/config.yaml
@@ -1,3 +1,0 @@
-build:
-  command: cargo build --target wasm32-wasip2 --release
-  component_path: ../target/wasm32-wasip2/release/nats_jetstream_handler.wasm
