--- crates/wash-runtime/tests/fixtures/nats-jetstream-handler/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "nats-jetstream-handler"
-edition = "2024"
-version = "0.1.0"
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-wit-bindgen = { workspace = true }
