--- crates/wash-runtime/tests/fixtures/nats-jetstream-handler/wit/world.wit
@@ -1,8 +1,0 @@
-package wasmcloud:nats-jetstream-handler-fixture;
-
-world nats-jetstream-handler-fixture {
-    import wasmcloud:nats/types@0.1.0;
-    import wasmcloud:nats/jetstream@0.1.0;
-    import wasmcloud:nats/kv@0.1.0;
-    export wasmcloud:nats/jetstream-handler@0.1.0;
-}
